# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "MyTodos API",
    "description": "Documentation for MyTodos"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/users/login": {
      "post": {
        "tags": [
          "Users"
        ],
        "description": "Login for a user",
        "produces": [
          "text/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Login"
            }
          }
        ],
        "responses": {
          "400, 500": {
            "description": ""
          }
        }
      }
    },
    "/users/signup": {
      "post": {
        "tags": [
          "Users"
        ],
        "description": "Sign up for new user",
        "produces": [
          "text/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/SignUp"
            }
          }
        ],
        "responses": {
          "200,400": {
            "description": ""
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "SignUp": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "kronos"
        },
        "email": {
          "type": "string",
          "example": "geo.elgeo@gmail.com"
        },
        "password": {
          "type": "string",
          "example": "123456"
        }
      },
      "required": [
        "name",
        "email",
        "password"
      ]
    },
    "Login": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "geo.elgeo@gmail.com"
        },
        "password": {
          "type": "string",
          "example": "123456"
        }
      },
      "required": [
        "email",
        "password"
      ]
    }
  }
}